backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
motor==3.6.0
pymongo==4.9.2
email-validator==2.2.0
pandas>=2.2,<2.3
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
reportlab==4.2.5
langchain==0.2.17
langchain-google-genai==1.0.10
